+++ lib/compat/wordpress-6.9/l10n.php
@@ -6,37 +6,12 @@
  * @package gutenberg
  */
 
-/**
- * Override core's wp_get_word_count_type() introduced in WordPress 6.2.
- * Originally, get_word_count_type() method of the WP_Locale class is executed,
- * but the process is simulated here.
- *
- * This function should not be backported to core.
- */
-if ( ! function_exists( 'wp_get_word_count_type' ) ) {
-	/**
-	 * Retrieves the word count type based on the locale.
-	 *
-	 * @return string Locale-specific word count type.
-	 */
-	function wp_get_word_count_type() {
-		$word_count_type = _x( 'words', 'Word count type. Do not translate!', 'gutenberg' );
-
-		// Check for valid types.
-		if ( 'characters_excluding_spaces' !== $word_count_type && 'characters_including_spaces' !== $word_count_type ) {
-			// Defaults to 'words'.
-			$word_count_type = 'words';
-		}
-		return $word_count_type;
-	}
-}
-
 if ( ! function_exists( 'wp_word_count' ) ) {
 	/**
-	 * Count words or characters in a provided text string.
+	 * Counts words or characters in a provided text string.
 	 *
 	 * @param string $text  Text to count elements in.
-	 * @param string $type The type of count. Accepts 'words', 'characters_excluding_spaces', or 'characters_including_spaces'.
+	 * @param string $type  The type of count. Accepts 'words', 'characters_excluding_spaces', or 'characters_including_spaces'.
 	 * @param array  $settings {
 	 *     Optional. Array of arguments used to overrides for settings.
 	 *
@@ -78,7 +53,7 @@ if ( ! function_exists( 'wp_word_count' ) ) {
 
 		$count = 0;
 
-		if ( ! $text ) {
+		if ( '' === trim( $text ) ) {
 			return $count;
 		}
 
@@ -128,12 +103,6 @@ if ( ! function_exists( 'wp_word_count' ) ) {
 		}
 
 		// Match with the selected type regular expression to count the items.
-		preg_match_all( $settings[ $type . '_regexp' ], $text, $matches );
-
-		if ( $matches ) {
-			return count( $matches[0] );
-		}
-
-		return $count;
+		return (int) preg_match_all( $settings[ $type . '_regexp' ], $text );
 	}
 }